+++ pkg/band/reference_test.go
@@ -245,10 +245,12 @@ func testBand(t *testing.T, band serializableBand, version ttnpb.PHYVersion) {
 }
 
 func TestBandDefinitions(t *testing.T) {
+	t.Parallel()
+
 	for name, versions := range band.All {
 		for version, band := range versions {
-			t.Parallel()
 			t.Run(fmt.Sprintf("%v/%v", name, version), func(t *testing.T) {
+				t.Parallel()
 				testBand(t, makeBand(band), version)
 			})
 		}