+++ pkg/gatewayserver/gatewayserver_test.go
@@ -899,12 +899,12 @@ func TestGatewayServer(t *testing.T) {
 													Lora: &ttnpb.LoRaDataRate{
 														SpreadingFactor: 7,
 														Bandwidth:       250000,
+														CodingRate:      "4/5",
 													},
 												},
 											},
-											CodingRate: "4/5",
-											Frequency:  867900000,
-											Timestamp:  100,
+											Frequency: 867900000,
+											Timestamp: 100,
 										},
 										RxMetadata: []*ttnpb.RxMetadata{
 											{
@@ -1058,12 +1058,12 @@ func TestGatewayServer(t *testing.T) {
 														Lora: &ttnpb.LoRaDataRate{
 															SpreadingFactor: 7,
 															Bandwidth:       250000,
+															CodingRate:      "4/5",
 														},
 													},
 												},
-												CodingRate: "4/5",
-												Frequency:  867900000,
-												Timestamp:  100,
+												Frequency: 867900000,
+												Timestamp: 100,
 											},
 											RxMetadata: []*ttnpb.RxMetadata{
 												{
@@ -1092,12 +1092,12 @@ func TestGatewayServer(t *testing.T) {
 														Lora: &ttnpb.LoRaDataRate{
 															SpreadingFactor: 7,
 															Bandwidth:       250000,
+															CodingRate:      "4/5",
 														},
 													},
 												},
-												CodingRate: "4/5",
-												Frequency:  867900000,
-												Timestamp:  101,
+												Frequency: 867900000,
+												Timestamp: 101,
 											},
 											RxMetadata: []*ttnpb.RxMetadata{
 												{
@@ -1118,12 +1118,12 @@ func TestGatewayServer(t *testing.T) {
 														Lora: &ttnpb.LoRaDataRate{
 															SpreadingFactor: 7,
 															Bandwidth:       250000,
+															CodingRate:      "4/5",
 														},
 													},
 												},
-												CodingRate: "4/5",
-												Frequency:  867900000,
-												Timestamp:  100,
+												Frequency: 867900000,
+												Timestamp: 100,
 											},
 											RxMetadata: []*ttnpb.RxMetadata{
 												{
@@ -1144,12 +1144,12 @@ func TestGatewayServer(t *testing.T) {
 														Lora: &ttnpb.LoRaDataRate{
 															SpreadingFactor: 7,
 															Bandwidth:       250000,
+															CodingRate:      "4/5",
 														},
 													},
 												},
-												CodingRate: "4/5",
-												Frequency:  867900000,
-												Timestamp:  100,
+												Frequency: 867900000,
+												Timestamp: 100,
 											},
 											RxMetadata: []*ttnpb.RxMetadata{
 												{
@@ -1212,12 +1212,12 @@ func TestGatewayServer(t *testing.T) {
 														Lora: &ttnpb.LoRaDataRate{
 															SpreadingFactor: 9,
 															Bandwidth:       125000,
+															CodingRate:      "4/5",
 														},
 													},
 												},
-												CodingRate: "4/5",
-												Frequency:  868500000,
-												Timestamp:  100,
+												Frequency: 868500000,
+												Timestamp: 100,
 											},
 											RxMetadata: []*ttnpb.RxMetadata{
 												{
@@ -1238,12 +1238,12 @@ func TestGatewayServer(t *testing.T) {
 														Lora: &ttnpb.LoRaDataRate{
 															SpreadingFactor: 7,
 															Bandwidth:       125000,
+															CodingRate:      "4/5",
 														},
 													},
 												},
-												CodingRate: "4/5",
-												Frequency:  868100000,
-												Timestamp:  200,
+												Frequency: 868100000,
+												Timestamp: 200,
 											},
 											RxMetadata: []*ttnpb.RxMetadata{
 												{
@@ -1264,12 +1264,12 @@ func TestGatewayServer(t *testing.T) {
 														Lora: &ttnpb.LoRaDataRate{
 															SpreadingFactor: 12,
 															Bandwidth:       125000,
+															CodingRate:      "4/5",
 														},
 													},
 												},
-												CodingRate: "4/5",
-												Frequency:  867700000,
-												Timestamp:  300,
+												Frequency: 867700000,
+												Timestamp: 300,
 											},
 											RxMetadata: []*ttnpb.RxMetadata{
 												{
@@ -1458,11 +1458,11 @@ func TestGatewayServer(t *testing.T) {
 													Lora: &ttnpb.LoRaDataRate{
 														SpreadingFactor: 12,
 														Bandwidth:       125000,
+														CodingRate:      "4/5",
 													},
 												},
 											},
-											CodingRate: "4/5",
-											Frequency:  869525000,
+											Frequency: 869525000,
 											Downlink: &ttnpb.TxSettings_Downlink{
 												TxPower: 10,
 											},