+++ pkg/gatewayserver/io/ws/lbslns/downstream_test.go
@@ -54,6 +54,7 @@ func TestFromDownlinkMessage(t *testing.T) {
 								Lora: &ttnpb.LoRaDataRate{
 									SpreadingFactor: 10,
 									Bandwidth:       125000,
+									CodingRate:      "4/5",
 								},
 							},
 						},
@@ -97,6 +98,7 @@ func TestFromDownlinkMessage(t *testing.T) {
 								Lora: &ttnpb.LoRaDataRate{
 									SpreadingFactor: 10,
 									Bandwidth:       125000,
+									CodingRate:      "4/5",
 								},
 							},
 						},
@@ -174,6 +176,7 @@ func TestToDownlinkMessage(t *testing.T) {
 								Lora: &ttnpb.LoRaDataRate{
 									SpreadingFactor: 10,
 									Bandwidth:       125000,
+									CodingRate:      "4/5",
 								},
 							},
 						},
@@ -209,6 +212,7 @@ func TestToDownlinkMessage(t *testing.T) {
 								Lora: &ttnpb.LoRaDataRate{
 									SpreadingFactor: 10,
 									Bandwidth:       125000,
+									CodingRate:      "4/5",
 								},
 							},
 						},