+++ pkg/packetbrokeragent/agent_test.go
@@ -151,11 +151,11 @@ func TestForwarder(t *testing.T) {
 									Lora: &ttnpb.LoRaDataRate{
 										SpreadingFactor: 7,
 										Bandwidth:       125000,
+										CodingRate:      "4/5",
 									},
 								},
 							},
-							CodingRate: "4/5",
-							Frequency:  869525000,
+							Frequency: 869525000,
 						},
 					},
 					BandId: "EU_863_870",
@@ -176,7 +176,6 @@ func TestForwarder(t *testing.T) {
 						ForwarderReceiveTime: test.Must(pbtypes.TimestampProto(time.Date(2020, time.March, 24, 12, 0, 0, 0, time.UTC))).(*pbtypes.Timestamp),
 						DataRate:             packetbroker.NewLoRaDataRate(7, 125000, "4/5"),
 						Frequency:            869525000,
-						CodingRate:           "4/5",
 						GatewayMetadata: &packetbroker.UplinkMessage_GatewayMetadata{
 							Teaser: &packetbroker.GatewayMetadataTeaser{
 								Value: &packetbroker.GatewayMetadataTeaser_Terrestrial_{
@@ -271,11 +270,11 @@ func TestForwarder(t *testing.T) {
 									Lora: &ttnpb.LoRaDataRate{
 										SpreadingFactor: 9,
 										Bandwidth:       125000,
+										CodingRate:      "4/5",
 									},
 								},
 							},
-							CodingRate: "4/5",
-							Frequency:  868300000,
+							Frequency: 868300000,
 						},
 					},
 					BandId: "EU_863_870",
@@ -296,7 +295,6 @@ func TestForwarder(t *testing.T) {
 						ForwarderReceiveTime: test.Must(pbtypes.TimestampProto(time.Date(2020, time.March, 24, 12, 0, 0, 0, time.UTC))).(*pbtypes.Timestamp),
 						DataRate:             packetbroker.NewLoRaDataRate(9, 125000, "4/5"),
 						Frequency:            868300000,
-						CodingRate:           "4/5",
 						GatewayMetadata: &packetbroker.UplinkMessage_GatewayMetadata{
 							Teaser: &packetbroker.GatewayMetadataTeaser{
 								Value: &packetbroker.GatewayMetadataTeaser_Terrestrial_{
@@ -722,11 +720,11 @@ func TestHomeNetwork(t *testing.T) {
 								Lora: &ttnpb.LoRaDataRate{
 									SpreadingFactor: 7,
 									Bandwidth:       125000,
+									CodingRate:      "4/5",
 								},
 							},
 						},
-						Frequency:  869525000,
-						CodingRate: "4/5",
+						Frequency: 869525000,
 					},
 				},
 			},
@@ -833,11 +831,11 @@ func TestHomeNetwork(t *testing.T) {
 								Lora: &ttnpb.LoRaDataRate{
 									SpreadingFactor: 9,
 									Bandwidth:       125000,
+									CodingRate:      "4/5",
 								},
 							},
 						},
-						Frequency:  869525000,
-						CodingRate: "4/5",
+						Frequency: 869525000,
 					},
 				},
 			},