+++ pkg/toa/toa_test.go
@@ -25,12 +25,11 @@ import (
 	"go.thethings.network/lorawan-stack/v3/pkg/util/test/assertions/should"
 )
 
-func buildLoRaDownlinkFromParameters(payloadSize int, frequency uint64, dataRate *ttnpb.DataRate, codingRate string) (downlink *ttnpb.DownlinkMessage, err error) {
+func buildLoRaDownlinkFromParameters(payloadSize int, frequency uint64, dataRate *ttnpb.DataRate) (downlink *ttnpb.DownlinkMessage, err error) {
 	payload := bytes.Repeat([]byte{0x0}, payloadSize)
 	scheduled := &ttnpb.TxSettings{
-		Frequency:  frequency,
-		DataRate:   dataRate,
-		CodingRate: codingRate,
+		Frequency: frequency,
+		DataRate:  dataRate,
 	}
 	downlink = &ttnpb.DownlinkMessage{
 		RawPayload: payload,
@@ -51,9 +50,10 @@ func TestInvalidLoRa(t *testing.T) {
 				Lora: &ttnpb.LoRaDataRate{
 					SpreadingFactor: 10,
 					Bandwidth:       125000,
+					CodingRate:      "1/9",
 				},
 			},
-		}, "1/9")
+		})
 		a.So(err, should.BeNil)
 		_, err = Compute(len(downlink.RawPayload), downlink.GetScheduled())
 		a.So(err, should.NotBeNil)
@@ -66,9 +66,10 @@ func TestInvalidLoRa(t *testing.T) {
 				Lora: &ttnpb.LoRaDataRate{
 					SpreadingFactor: 0,
 					Bandwidth:       125000,
+					CodingRate:      "4/5",
 				},
 			},
-		}, "4/5")
+		})
 		a.So(err, should.BeNil)
 		_, err = Compute(len(downlink.RawPayload), downlink.GetScheduled())
 		a.So(err, should.NotBeNil)
@@ -81,9 +82,10 @@ func TestInvalidLoRa(t *testing.T) {
 				Lora: &ttnpb.LoRaDataRate{
 					SpreadingFactor: 7,
 					Bandwidth:       0,
+					CodingRate:      "4/5",
 				},
 			},
-		}, "4/5")
+		})
 		a.So(err, should.BeNil)
 		_, err = Compute(len(downlink.RawPayload), downlink.GetScheduled())
 		a.So(err, should.NotBeNil)
@@ -97,32 +99,56 @@ func TestDifferentLoRaSFs(t *testing.T) {
 		expected uint
 	}{
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 7, Bandwidth: 125000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 7,
+				Bandwidth:       125000,
+				CodingRate:      "4/5",
+			}}},
 			expected: 41216,
 		},
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 8, Bandwidth: 125000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 8,
+				Bandwidth:       125000,
+				CodingRate:      "4/5",
+			}}},
 			expected: 72192,
 		},
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 9, Bandwidth: 125000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 9,
+				Bandwidth:       125000,
+				CodingRate:      "4/5",
+			}}},
 			expected: 144384,
 		},
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 10, Bandwidth: 125000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 10,
+				Bandwidth:       125000,
+				CodingRate:      "4/5",
+			}}},
 			expected: 288768,
 		},
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 11, Bandwidth: 125000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 11,
+				Bandwidth:       125000,
+				CodingRate:      "4/5",
+			}}},
 			expected: 577536,
 		},
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 12, Bandwidth: 125000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 12,
+				Bandwidth:       125000,
+				CodingRate:      "4/5",
+			}}},
 			expected: 991232,
 		},
 	}
 	for _, tt := range sfTests {
-		dl, err := buildLoRaDownlinkFromParameters(10, 868100000, tt.dr, "4/5")
+		dl, err := buildLoRaDownlinkFromParameters(10, 868100000, tt.dr)
 		a.So(err, should.BeNil)
 		toa, err := Compute(len(dl.RawPayload), dl.GetScheduled())
 		a.So(err, should.BeNil)
@@ -137,20 +163,32 @@ func TestDifferentLoRaBWs(t *testing.T) {
 		expected uint
 	}{
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 7, Bandwidth: 125000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 7,
+				Bandwidth:       125000,
+				CodingRate:      "4/5",
+			}}},
 			expected: 41216,
 		},
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 7, Bandwidth: 250000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 7,
+				Bandwidth:       250000,
+				CodingRate:      "4/5",
+			}}},
 			expected: 20608,
 		},
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 7, Bandwidth: 500000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 7,
+				Bandwidth:       500000,
+				CodingRate:      "4/5",
+			}}},
 			expected: 10304,
 		},
 	}
 	for _, tt := range bwTests {
-		dl, err := buildLoRaDownlinkFromParameters(10, 868100000, tt.dr, "4/5")
+		dl, err := buildLoRaDownlinkFromParameters(10, 868100000, tt.dr)
 		a.So(err, should.BeNil)
 		toa, err := Compute(len(dl.RawPayload), dl.GetScheduled())
 		a.So(err, should.BeNil)
@@ -167,7 +205,11 @@ func TestDifferentLoRaCRs(t *testing.T) {
 		"4/8": 53504,
 	}
 	for cr, us := range crTests {
-		dl, err := buildLoRaDownlinkFromParameters(10, 868100000, &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 7, Bandwidth: 125000}}}, cr)
+		dl, err := buildLoRaDownlinkFromParameters(10, 868100000, &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+			SpreadingFactor: 7,
+			Bandwidth:       125000,
+			CodingRate:      cr,
+		}}})
 		a.So(err, should.BeNil)
 		toa, err := Compute(len(dl.RawPayload), dl.GetScheduled())
 		a.So(err, should.BeNil)
@@ -187,7 +229,11 @@ func TestDifferentLoRaPayloadSizes(t *testing.T) {
 		19: 51456,
 	}
 	for size, us := range plTests {
-		dl, err := buildLoRaDownlinkFromParameters(size, 868100000, &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 7, Bandwidth: 125000}}}, "4/5")
+		dl, err := buildLoRaDownlinkFromParameters(size, 868100000, &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+			SpreadingFactor: 7,
+			Bandwidth:       125000,
+			CodingRate:      "4/5",
+		}}})
 		a.So(err, should.BeNil)
 		toa, err := Compute(len(dl.RawPayload), dl.GetScheduled())
 		a.So(err, should.BeNil)
@@ -235,9 +281,10 @@ func TestInvalidLoRa2400(t *testing.T) {
 				Lora: &ttnpb.LoRaDataRate{
 					SpreadingFactor: 10,
 					Bandwidth:       812000,
+					CodingRate:      "1/9LI",
 				},
 			},
-		}, "1/9LI")
+		})
 		a.So(err, should.BeNil)
 		_, err = Compute(len(downlink.RawPayload), downlink.GetScheduled())
 		a.So(err, should.NotBeNil)
@@ -250,9 +297,10 @@ func TestInvalidLoRa2400(t *testing.T) {
 				Lora: &ttnpb.LoRaDataRate{
 					SpreadingFactor: 0,
 					Bandwidth:       812000,
+					CodingRate:      "4/5LI",
 				},
 			},
-		}, "4/5LI")
+		})
 		a.So(err, should.BeNil)
 		_, err = Compute(len(downlink.RawPayload), downlink.GetScheduled())
 		a.So(err, should.NotBeNil)
@@ -265,9 +313,10 @@ func TestInvalidLoRa2400(t *testing.T) {
 				Lora: &ttnpb.LoRaDataRate{
 					SpreadingFactor: 7,
 					Bandwidth:       0,
+					CodingRate:      "4/5LI",
 				},
 			},
-		}, "4/5LI")
+		})
 		a.So(err, should.BeNil)
 		_, err = Compute(len(downlink.RawPayload), downlink.GetScheduled())
 		a.So(err, should.NotBeNil)
@@ -281,40 +330,72 @@ func TestDifferentLoRa2400SFs(t *testing.T) {
 		expected time.Duration
 	}{
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 5, Bandwidth: 812000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 5,
+				Bandwidth:       812000,
+				CodingRate:      "4/5LI",
+			}}},
 			expected: 1665000,
 		},
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 6, Bandwidth: 812000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 6,
+				Bandwidth:       812000,
+				CodingRate:      "4/5LI",
+			}}},
 			expected: 3093600,
 		},
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 7, Bandwidth: 812000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 7,
+				Bandwidth:       812000,
+				CodingRate:      "4/5LI",
+			}}},
 			expected: 5556700,
 		},
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 8, Bandwidth: 812000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 8,
+				Bandwidth:       812000,
+				CodingRate:      "4/5LI",
+			}}},
 			expected: 10482800,
 		},
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 9, Bandwidth: 812000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 9,
+				Bandwidth:       812000,
+				CodingRate:      "4/5LI",
+			}}},
 			expected: 19073900,
 		},
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 10, Bandwidth: 812000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 10,
+				Bandwidth:       812000,
+				CodingRate:      "4/5LI",
+			}}},
 			expected: 36886700,
 		},
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 11, Bandwidth: 812000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 11,
+				Bandwidth:       812000,
+				CodingRate:      "4/5LI",
+			}}},
 			expected: 73773400,
 		},
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 12, Bandwidth: 812000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 12,
+				Bandwidth:       812000,
+				CodingRate:      "4/5LI",
+			}}},
 			expected: 142502500,
 		},
 	}
 	for _, tt := range sfTests {
-		dl, err := buildLoRaDownlinkFromParameters(10, 2422000000, tt.dr, "4/5LI")
+		dl, err := buildLoRaDownlinkFromParameters(10, 2422000000, tt.dr)
 		a.So(err, should.BeNil)
 		toa, err := Compute(len(dl.RawPayload), dl.GetScheduled())
 		a.So(err, should.BeNil)
@@ -329,24 +410,40 @@ func TestDifferentLoRa2400BWs(t *testing.T) {
 		expected time.Duration
 	}{
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 7, Bandwidth: 203000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 7,
+				Bandwidth:       203000,
+				CodingRate:      "4/5LI",
+			}}},
 			expected: 22226600,
 		},
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 7, Bandwidth: 406000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 7,
+				Bandwidth:       406000,
+				CodingRate:      "4/5LI",
+			}}},
 			expected: 11113300,
 		},
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 7, Bandwidth: 812000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 7,
+				Bandwidth:       812000,
+				CodingRate:      "4/5LI",
+			}}},
 			expected: 5556700,
 		},
 		{
-			dr:       &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{SpreadingFactor: 7, Bandwidth: 1625000}}},
+			dr: &ttnpb.DataRate{Modulation: &ttnpb.DataRate_Lora{Lora: &ttnpb.LoRaDataRate{
+				SpreadingFactor: 7,
+				Bandwidth:       1625000,
+				CodingRate:      "4/5LI",
+			}}},
 			expected: 2776600,
 		},
 	}
 	for _, tt := range bwTests {
-		dl, err := buildLoRaDownlinkFromParameters(10, 2422000000, tt.dr, "4/5LI")
+		dl, err := buildLoRaDownlinkFromParameters(10, 2422000000, tt.dr)
 		a.So(err, should.BeNil)
 		toa, err := Compute(len(dl.RawPayload), dl.GetScheduled())
 		a.So(err, should.BeNil)
@@ -368,9 +465,10 @@ func TestDifferentLoRa2400CRs(t *testing.T) {
 				Lora: &ttnpb.LoRaDataRate{
 					SpreadingFactor: 7,
 					Bandwidth:       812000,
+					CodingRate:      cr,
 				},
 			},
-		}, cr)
+		})
 		a.So(err, should.BeNil)
 		toa, err := Compute(len(dl.RawPayload), dl.GetScheduled())
 		a.So(err, should.BeNil)
@@ -394,9 +492,10 @@ func TestDifferentLoRa2400PayloadSizes(t *testing.T) {
 				Lora: &ttnpb.LoRaDataRate{
 					SpreadingFactor: 12,
 					Bandwidth:       812000,
+					CodingRate:      "4/5LI",
 				},
 			},
-		}, "4/5LI")
+		})
 		a.So(err, should.BeNil)
 		toa, err := Compute(len(dl.RawPayload), dl.GetScheduled())
 		a.So(err, should.BeNil)
@@ -416,9 +515,10 @@ func TestDifferentLoRa2400CRCs(t *testing.T) {
 				Lora: &ttnpb.LoRaDataRate{
 					SpreadingFactor: 7,
 					Bandwidth:       812000,
+					CodingRate:      "4/5LI",
 				},
 			},
-		}, "4/5LI")
+		})
 		dl.GetScheduled().EnableCrc = crc
 		a.So(err, should.BeNil)
 		toa, err := Compute(len(dl.RawPayload), dl.GetScheduled())