+++ pkg/ttnpb/lorawan.pb.paths.fm.go
@@ -240,7 +240,6 @@ var DataRateFieldPathsTopLevel = []string{
 	"modulation",
 }
 var TxSettingsFieldPathsNested = []string{
-	"coding_rate",
 	"concentrator_timestamp",
 	"data_rate",
 	"data_rate.modulation",
@@ -265,7 +264,6 @@ var TxSettingsFieldPathsNested = []string{
 }
 
 var TxSettingsFieldPathsTopLevel = []string{
-	"coding_rate",
 	"concentrator_timestamp",
 	"data_rate",
 	"downlink",