+++ pkg/ttnpb/lorawan.pb.validate.go
@@ -1692,8 +1692,6 @@ func (m *TxSettings) ValidateFields(paths ...string) error {
 				}
 			}
 
-		case "coding_rate":
-			// no validation rules for CodingRate
 		case "frequency":
 			// no validation rules for Frequency
 		case "enable_crc":